LLDB handles all

+++ dumpdecrypted.py
@@ -4,17 +4,26 @@ import lldb
 import commands
 import optparse
 import shlex
+import struct
 
 def create_dumpdecrypted_options():
     usage = "usage: %prog <options>"
     parser = optparse.OptionParser(prog='dumpdecrypted', usage=usage)
     parser.add_option('-i', '--image', type='string', dest='image', help='dumpdecrypted image name')
-    parser.add_option('-s', '--cryptsize', type='int', dest='size', help='crypted size')
-    parser.add_option('-f', '--cryptoff', type='int', dest='cryptoff', help='crypted offset')
     parser.add_option('-o', '--output', type='string', dest='output', help='Where to save the output file')
     return parser
 
 def dumpdecrypted(debugger, command, result, internal_dict):
+    # MachO Magic Constants
+    MH_MAGIC = 0xFEEDFACE
+    MH_CIGAM = 0xCEFAEDFE
+    MH_MAGIC_64 = 0xFEEDFACF
+    MH_CIGAM_64 = 0xCFFAEDFE
+    LC_ENCRYPTION_INFO_64 = 0x2C
+    LC_ENCRYPTION_INFO = 0x21
+    MACHO_HEADER_SIZE = 28
+    MACHO_HEADER_64_SIZE = 32
+    
     command_list = shlex.split(command)
     
     # Parse the options
@@ -24,23 +33,68 @@ def dumpdecrypted(debugger, command, result, internal_dict):
     except:
         result.SetError("Option parsing failed")
         return
-
+    
     target = debugger.GetSelectedTarget()
     for module in target.modules:
         if module.file.basename == options.image:
-            image_load_address = module.GetObjectFileHeaderAddress().load_addr
+            image_load_address = module.GetObjectFileHeaderAddress().GetLoadAddress(target)
             print("INFO: image `%s' loaded at <0x%016X>" % (options.image, image_load_address))
-            ci = debugger.GetCommandInterpreter()
-            res = lldb.SBCommandReturnObject()
-            ci.HandleCommand("memory read --force --outfile %s.bin --binary --count %d 0x%016X" % (options.output, options.size, image_load_address + options.cryptoff), res)
-            if res.Succeeded():
-                print("INFO: %d bytes read as binary" % (options.size))
-                print("INFO: convert binary file via dd...")
-                result = commands.getoutput('dd seek=%d bs=1 conv=notrunc if=%s.bin of=%s' % (options.cryptoff, options.output, options.output))
-                commands.getoutput('rm -f %s.bin' % (options.output))
-                print(result)
+
+            # read magic
+            mach_header_size = MACHO_HEADER_SIZE
+            encryption_info_cmd = LC_ENCRYPTION_INFO
+            encryption_info_description = "LC_ENCRYPTION_INFO"
+            
+            process = target.GetProcess();
+            read_res = lldb.SBError()
+            macho_memory = process.ReadMemory(image_load_address, 24, read_res)
+            if read_res.Success():
+                # endian
+                magic, = struct.unpack("<I", macho_memory[0:4])
+                if magic == MH_MAGIC or magic == MH_MAGIC_64:
+                    prefix = "<"
+                elif magic == MH_CIGAM or magic == MH_CIGAM_64:
+                    prefix = ">"
+                else:
+                    raise Exception("magic wrong")
+
+                # 64 bit
+                if magic == MH_MAGIC_64 or magic == MH_CIGAM_64:
+                    mach_header_size = MACHO_HEADER_64_SIZE
+                    encryption_info_cmd = LC_ENCRYPTION_INFO_64
+                    encryption_info_description = "LC_ENCRYPTION_INFO_64"
+
+                # read load_command
+                ncmds,sizecmds = struct.unpack(prefix + "2I", macho_memory[16:24])
+                macho_memory = process.ReadMemory(image_load_address, sizecmds, read_res)
+                cryptoffset = 0
+                cryptsize = 0
+                cryptid = 0
+                load_command_start = mach_header_size
+                for i in range(ncmds):
+                    cmd, cmdsize = struct.unpack(prefix + "2I", macho_memory[load_command_start:load_command_start+8])
+                    if cmd == encryption_info_cmd:
+                        print("INFO: found %s" % (encryption_info_description))
+                        load_command_start += 8
+                        cryptoffset, cryptsize, cryptid = struct.unpack(prefix + "3I", macho_memory[load_command_start:load_command_start+12])
+                        print("INFO: cryptoffset: %d\n      cryptsize:%d\n      cryptid: %d" % (cryptoffset, cryptsize, cryptid))
+                        break
+                    else:
+                        load_command_start += cmdsize
+
+                ci = debugger.GetCommandInterpreter()
+                res = lldb.SBCommandReturnObject()
+                ci.HandleCommand("memory read --force --outfile %s.bin --binary --count %d 0x%016X" % (options.output, cryptsize, image_load_address + cryptoffset), res)
+                if res.Succeeded():
+                    print("INFO: %d bytes read as binary" % (cryptsize))
+                    print("INFO: convert binary file via dd...")
+                    result = commands.getoutput('dd seek=%d bs=1 conv=notrunc if=%s.bin of=%s' % (cryptoffset, options.output, options.output))
+                    commands.getoutput('rm -f %s.bin' % (options.output))
+                    print(result)
+                else:
+                    print(res)
             else:
-                print(res)
+                print(read_res)
             break
 
 def __lldb_init_module(debugger, internal_dict):
